build: move solution to src and align MSBuild paths
- Use src/splayer.sln as the canonical solution; remove root splayer.sln
- Replace $(SolutionDir)src\ with $(SolutionDir) and $(SolutionDir)out\ with $(SolutionDir)..\out\ across vcxproj/props
- Point BuildScript and docs to src/splayer.sln; update path-fix scripts
- Fix Thirdparty zlib/libpng props and MediaTree_Test includes
- Consolidate root notes under docs/root-notes; relocate install script to script/

Made-with: Cursor

+++ src/Source/subtitles/libssf/Subtitle.cpp
@@ -302,8 +302,8 @@ namespace ssf
 				CStringW direction = (*pDef)[L"direction"].IsValue() ? (*pDef)[L"direction"] : L"fw";
 				if(direction == L"fwbw" || direction == L"bwfw") t *= 2;
 
-				float n;
-				t = modf(t, &n);
+				double n;
+				t = (float)modf((double)t, &n);
 
 				if(direction == L"bw" 
 				|| direction == L"fwbw" && ((int)n & 1)
